Resolved conflicts:
* Fix for intermittent test failure in TestPlotView
  had been implemented in separate branches;
  accepted most-commented version of the fix.
* Swing thread usage had been added on one
  branch for TestPlotViewRole, and TestNG data
  provider usage had been added on another.
  Merged both approaches.

+++ fastPlotViews/src/main/java/gov/nasa/arc/mct/fastplot/settings/controls/PlotSettingsAxisGroup.java
@@ -165,7 +165,7 @@ public abstract class PlotSettingsAxisGroup extends PlotSettingsPanel implements
 			minControls.updateCurrent(getActualMinimum(view));
 			maxControls.updateCurrent(getActualMaximum(view));
 			minControls.updateAuto((double) view.getCurrentMCTTime());
-			maxControls.updateAuto((double) getValue(minControls) + spanControls.getSpanValue());
+			maxControls.updateAuto(getValue(minControls) + spanControls.getSpanValue());
 			if (!maxControls.auto.isSelected()) {
 				spanControls.setSpanValue(getValue(maxControls) - getValue(minControls));
 			}
